Split tag_analysis from doc_analysis method
